requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Limiter==3.12
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
rich==13.9.4
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
